backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
motor==3.7.1
pydantic==2.12.5
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
